+++ ving/redis.mjs
@@ -11,7 +11,28 @@ export const spawnRedis = () => {
     if (!redisUrl) {
         throw new Error('VING_REDIS environment variable is not set');
     }
-    return new Redis(redisUrl, { maxRetriesPerRequest: null });
+    const redisConfig = new URL(redisUrl);
+    const redisNode = {
+        port: parseInt(redisConfig.port || 6379),
+        host: redisConfig.hostname,
+    };
+    const redisOptions = { maxRetriesPerRequest: null, keyPrefix: process.env.NODE_ENV };
+    if (redisConfig.password) {
+        redisOptions.password = redisConfig.password;
+    }
+    if (redisConfig.username) {
+        redisOptions.username = redisConfig.username;
+    }
+    if (redisConfig.protocol === 'rediss:') {
+        redisOptions.tls = {};
+    }
+    if (redisConfig.searchParams.get('cluster') === 'yes') {
+        return new Redis.Cluster([redisNode], {
+            slotsRefreshTimeout: 20000,
+            redisOptions,
+        });
+    }
+    return new Redis(redisNode.port, redisNode.host, redisOptions);
 }
 
 let redis = undefined;